Sound effect skeleton
Uploaded some skeleton code for how sound will work in the game.

SoundManager is another Singleton that contains pointers to SoundEffect
objects which comprises all sound effects that will play during the
game.

Each SoundEffect is created when the manager is created by loading in
the sound file from disk.  No idea how to do this in Ogre, yet.

NYI: When a sound needs to be played, the object to play the sound will
need only to include the SoundManager, then choose which sound to play
and invoke its play() function.