Input
KeyMap.h is a header file that should contain only #define's of all of
our controls followed by a character.

This will help if we ever want to change the controls as well as provide
an easy way for us all to see what the controls are.

The base application should check if the input was designated to turn
off the sound only.  If this is not the case, it should be passed to
room->input(keyPressed)